requirements.txt
scikit-image==0.17.2
python_speech_features==0.6
pyworld==0.3.3
pyyaml==6.0
pytorch-lightning==1.5.10
imageio==2.15.0
scipy==1.5.4
pyworld==0.3.3
opencv-python==3.4.17.63